Ajustes a los DAOS y entidades en general Retirar los transaction del repository, pendiente categoria y carro

+++ app/src/main/java/com/swbvelasquez/nekoexpress/core/util/Functions.kt
@@ -40,7 +40,7 @@ object Functions {
         return NekoApplication.gson.fromJson(this,type)
     }
 
-    fun getCompleteCategoryDto(index:Int,category: String):CategoryDto = CategoryDto(index+1,category, getCategoryImage(category))
+    fun getCompleteCategoryDto(index:Long,category: String):CategoryDto = CategoryDto(index+1,category, getCategoryImage(category))
 
     private fun getCategoryImage(category: String):String{
         return when(category){
