+++ app/src/main/java/com/swbvelasquez/nekoexpress/data/database/entity/RatingEntity.kt
@@ -8,9 +8,9 @@ import com.swbvelasquez.nekoexpress.domain.model.RatingModel
 @Entity(tableName = "RatingTable")
 data class RatingEntity(
     @PrimaryKey(autoGenerate = false)
-    @ColumnInfo(name = "productId") val productId:Int=-1,
+    @ColumnInfo(name = "productId") val productId:Long=-1,
     @ColumnInfo(name = "rate") val rate:Double,
     @ColumnInfo(name = "count") val count:Int
 )
 
-fun RatingModel.toRatingEntity(productId:Int) = RatingEntity(productId = productId, rate = rate, count = count)
+fun RatingModel.toRatingEntity(productId:Long) = RatingEntity(productId = productId, rate = rate, count = count)
